types: change dev_t to uint32_t
* api/dev.h: header with definitions of _MAJOR, _MINOR and _MKDEV
* sys/sysmacros.h: header including api/dev.h and defining major, minor
  and mkdev
* dev.h: redefine macros for block devices to have more user-friendly
  numbers in minor
* file_system.c: fix bug - using inode.dev instead of rdev as mount
  device

+++ include/kernel/api/types.h
@@ -9,7 +9,7 @@ typedef uint32_t    blkcnt_t;
 typedef size_t      blksize_t;
 typedef uint32_t    clock_t;
 typedef uint32_t    clockid_t;
-typedef uint16_t    dev_t;
+typedef uint32_t    dev_t;
 typedef uint32_t    fsblkcnt_t;
 typedef uint32_t    fsfilcnt_t;
 typedef int         gid_t;
